+++ examples/resources/groundcover_apikey/import.sh
@@ -1,0 +1,1 @@
+terraform import groundcover_apikey.example "<id>"