+++ examples/resources/groundcover_connected_app/import.sh
@@ -1,0 +1,1 @@
+terraform import groundcover_connected_app.example "<id>"