+++ examples/resources/groundcover_logspipeline/import.sh
@@ -1,0 +1,2 @@
+# Logs pipeline is a singleton resource. The import ID value is ignored.
+terraform import groundcover_logspipeline.example any