+++ examples/resources/groundcover_notification_route/import.sh
@@ -1,0 +1,1 @@
+terraform import groundcover_notification_route.example "<id>"