+++ examples/resources/groundcover_tracespipeline/import.sh
@@ -1,0 +1,2 @@
+# Traces pipeline is a singleton resource. The import ID value is ignored.
+terraform import groundcover_tracespipeline.example any